dbt-requirements.txt
dbt-core
dbt-duckdb
requests
airflow-provider-duckdb
duckdb